+++ src/main/java/com/kelseyde/calvin/utils/notation/FEN.java
@@ -6,125 +6,144 @@ import com.kelseyde.calvin.board.Bits.Square;
 import com.kelseyde.calvin.uci.UCI;
 
 import java.util.*;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-public record FEN(String value) {
+/**
+ * FEN (Forsyth-Edwards Notation) is the standard notation for describing the state of a chess position. It encodes into a
+ * single string the position of the pieces, the side to move, castling rights, en passant target square, halfmove clock,
+ * and fullmove number.
+ * <p>
+ * This class provides methods to parse a FEN string into a {@link Board} object and to convert a {@link Board} object
+ * back into a FEN string.
+ */
+public record FEN(String value,
+                  long[] bitboards,
+                  Piece[] pieces,
+                  int enPassantFile,
+                  boolean whiteToMove,
+                  int castleRights,
+                  int halfMoveClock,
+                  int fullMoveNumber) {
 
     public static final String STARTPOS = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
 
-    public static Board startpos() {
-        return toBoard(STARTPOS);
+    public static FEN startpos() {
+        return FEN.parse(FEN.STARTPOS);
     }
 
+    /**
+     * Parses a FEN string.
+     * Checks that the string is a valid FEN string, and throws an {@link InvalidFenException} if it is not.
+     */
     public static FEN parse(String fen) throws InvalidFenException {
 
         if (fen == null || fen.isEmpty())
-            throw new InvalidFenException("FEN string is null or empty");
+            throw new InvalidFenException("FEN string is null or empty", fen);
 
-        String[] fenParts = fen.split(" ");
+        String[] parts = fen.split(" ");
 
-        if (fenParts.length < 4 || fenParts.length > 6)
-            throw new InvalidFenException("FEN string has an invalid number of parts");
+        if (parts.length < 4 || parts.length > 6)
+            throw new InvalidFenException("FEN string has an invalid number of parts", fen);
 
-        String[] ranks = fenParts[0].split("/");
+        String board = parts[0];
 
-        if (ranks.length != 8)
-            throw new InvalidFenException("FEN string has an invalid number of ranks");
+        int slashCount = (int) board.chars().filter(ch -> ch == '/').count();
+        if (slashCount != 7)
+            throw new InvalidFenException("Invalid number of ranks", fen);
 
+        String[] ranks = board.split("/");
+        for (String rank : ranks) {
 
+            if (FenPatterns.RANK.matcher(rank).matches())
+                throw new InvalidFenException("Consecutive numbers in rank are not allowed", fen);
 
-        return null;
-    }
-
-    public static Board toBoard(String fen) {
-
-            String[] parts = fen.split(" ");
-            String[] files = parts[0].split("/");
-
-            long whitePawns = 0L;
-            long whiteKnights = 0L;
-            long whiteBishops = 0L;
-            long whiteRooks = 0L;
-            long whiteQueens = 0L;
-            long whiteKing = 0L;
-            long blackPawns = 0L;
-            long blackKnights = 0L;
-            long blackBishops = 0L;
-            long blackRooks = 0L;
-            long blackQueens = 0L;
-            long blackKing = 0L;
-
-            List<List<String>> rankFileHash = Arrays.stream(files)
-                    .map(file -> Arrays.stream(file.split(""))
-                            .flatMap(FEN::parseSquare)
-                            .toList())
-                    .toList();
-            Collections.reverse(rankFileHash);
-
-            for (int rankIndex = 0; rankIndex < rankFileHash.size(); rankIndex++) {
-                List<String> rank = rankFileHash.get(rankIndex);
-                for (int fileIndex = 0; fileIndex < rank.size(); fileIndex++) {
-                    int square = Square.of(rankIndex, fileIndex);
-                    String squareValue = rank.get(fileIndex);
-                    long squareBB = Bits.of(square);
-                    switch (squareValue) {
-                        case "P" -> whitePawns |= squareBB;
-                        case "N" -> whiteKnights |= squareBB;
-                        case "B" -> whiteBishops |= squareBB;
-                        case "R" -> whiteRooks |= squareBB;
-                        case "Q" -> whiteQueens |= squareBB;
-                        case "K" -> whiteKing |= squareBB;
-                        case "p" -> blackPawns |= squareBB;
-                        case "n" -> blackKnights |= squareBB;
-                        case "b" -> blackBishops |= squareBB;
-                        case "r" -> blackRooks |= squareBB;
-                        case "q" -> blackQueens |= squareBB;
-                        case "k" -> blackKing |= squareBB;
-                    }
-                }
+            int squareCount = 0;
+            for (int i = 0; i < rank.length(); i++) {
+                char c = rank.charAt(i);
+                if (Character.isDigit(c))
+                    squareCount += Character.getNumericValue(c);
+                 else if (Character.isLetter(c))
+                    squareCount += 1;
             }
 
-            long pawns = whitePawns | blackPawns;
-            long knight = whiteKnights | blackKnights;
-            long bishops = whiteBishops | blackBishops;
-            long rooks = whiteRooks | blackRooks;
-            long queens = whiteQueens | blackQueens;
-            long king = whiteKing | blackKing;
-            long whitePieces = whitePawns | whiteKnights | whiteBishops | whiteRooks | whiteQueens | whiteKing;
-            long blackPieces = blackPawns | blackKnights | blackBishops | blackRooks | blackQueens | blackKing;
+            if (squareCount != 8) {
+                if (squareCount > 8)
+                    throw new InvalidFenException("Rank has too many pieces", fen);
+                else
+                    throw new InvalidFenException("Rank has too few pieces", fen);
+            }
+        }
 
-            boolean whiteToMove = parseSideToMove(parts[1]);
-            int castlingRights = parseCastlingRights(parts[2], whiteRooks, blackRooks, Bits.next(whiteKing), Bits.next(blackKing));
-            int enPassantFile = parseEnPassantFile(parts[3]);
-            int fiftyMoveCounter = parts.length > 4 ? parseFiftyMoveCounter(parts[4]) : 0;
-            // This implementation does not require the full move counter (parts[5]).
+        List<String> kings = parseKings(board);
+        if (kings.size() != 2)
+            throw new InvalidFenException("Invalid number of kings", fen);
 
-            Board board = new Board();
-            board.setBitboards(new long[Piece.COUNT + 2]);
-            board.setPawns(pawns);
-            board.setKnights(knight);
-            board.setBishops(bishops);
-            board.setRooks(rooks);
-            board.setQueens(queens);
-            board.setKings(king);
-            board.setWhitePieces(whitePieces);
-            board.setBlackPieces(blackPieces);
-            board.setPieces(calculatePieceList(board));
-            board.setWhite(whiteToMove);
-            board.getState().setRights(castlingRights);
-            board.getState().setEnPassantFile(enPassantFile);
-            board.getState().setHalfMoveClock(fiftyMoveCounter);
-            board.getState().setKey(Key.generateKey(board));
-            board.getState().setPawnKey(Key.generatePawnKey(board));
-            board.getState().setNonPawnKeys(Key.generateNonPawnKeys(board));
+        if (!kings.contains("K"))
+            throw new InvalidFenException("Missing white king", fen);
 
-            return board;
+        if (!kings.contains("k"))
+            throw new InvalidFenException("Missing black king", fen);
 
+        if (!FenPatterns.BOARD.matcher(board).matches())
+            throw new InvalidFenException("Invalid board representation", fen);
+
+        String turn = parts[1];
+        if (!FenPatterns.TURN.matcher(turn).matches())
+            throw new InvalidFenException("Invalid turn: " + turn, fen);
+
+        String castle = parts[2];
+        if (!FenPatterns.CASTLE.matcher(castle).matches())
+            throw new InvalidFenException("Invalid castling rights: " + castle, fen);
+
+        String enPassant = parts[3];
+        if (!FenPatterns.EN_PASSANT.matcher(enPassant).matches())
+            throw new InvalidFenException("Invalid en passant square: " + enPassant, fen);
+
+        String halfMove = parts.length > 4 ? parts[4] : "0";
+        if (!FenPatterns.HALF_MOVE.matcher(halfMove).matches())
+            throw new InvalidFenException("Invalid half move clock: " + halfMove, fen);
+
+        String fullMove = parts.length > 5 ? parts[5] : "0";
+        if (!FenPatterns.FULL_MOVE.matcher(fullMove).matches())
+            throw new InvalidFenException("Invalid full move number: " + fullMove, fen);
+
+        long[] bitboards = new long[Piece.COUNT + 2];
+        Piece[] pieces = new Piece[Square.COUNT];
+        calculatePiecePositions(board, bitboards, pieces);
+        int enPassantFile = parseEnPassantFile(enPassant);
+        boolean whiteToMove = turn.equals("w");
+        int castleRights = parseCastlingRights(castle, bitboards);
+        int halfMoveClock = Integer.parseInt(halfMove);
+        int fullMoveNumber = Integer.parseInt(fullMove);
+
+        return new FEN(fen, bitboards, pieces, enPassantFile, whiteToMove, castleRights, halfMoveClock, fullMoveNumber);
     }
 
-    public static String toFEN(Board board) {
+    /**
+     * Converts this {@link FEN} to a {@link Board}.
+     */
+    public Board toBoard() {
+        Board board = new Board();
+        board.setBitboards(bitboards);
+        board.setPieces(pieces);
+        board.setWhite(whiteToMove);
+        board.getState().setRights(castleRights);
+        board.getState().setHalfMoveClock(halfMoveClock);
+        board.getState().setFullMoveNumber(fullMoveNumber);
+        board.getState().setEnPassantFile(enPassantFile);
+        board.getState().setKey(Key.generateKey(board));
+        board.getState().setPawnKey(Key.generatePawnKey(board));
+        board.getState().setNonPawnKeys(Key.generateNonPawnKeys(board));
+        return board;
+    }
+
+    /**
+     * Converts a {@link Board} to a {@link FEN}.
+     */
+    public static FEN fromBoard(Board board) {
         try {
             StringBuilder sb = new StringBuilder();
 
@@ -147,12 +166,10 @@ public record FEN(String value) {
                         emptySquares++;
                     }
                 }
-                if (emptySquares != 0) {
+                if (emptySquares != 0)
                     sb.append(emptySquares);
-                }
-                if (rank > 0) {
+                if (rank > 0)
                     sb.append('/');
-                }
             }
 
             String whiteToMove = toSideToMove(board.isWhite());
@@ -170,28 +187,68 @@ public record FEN(String value) {
             String fullMoveNumber = toFullMoveCounter(board.getPly());
             sb.append(" ").append(fullMoveNumber);
 
-            return sb.toString();
+            return FEN.parse(sb.toString());
         } catch (Exception e) {
-            throw new IllegalArgumentException(board.toString(), e);
+            throw new InvalidFenException(board.toString(), e);
         }
     }
 
-    private static boolean parseSideToMove(String sideToMove) {
-        return switch (sideToMove) {
-            case "w" -> true;
-            case "b" -> false;
-            default -> throw new InvalidFenException("Invalid side to move! " + sideToMove);
-        };
+    @Override
+    public String toString() {
+        return value;
     }
 
-    private static String toSideToMove(boolean sideToMove) {
-        return sideToMove ? "w" : "b";
-    }
+    private static void calculatePiecePositions(String boardString, long[] bitboards, Piece[] pieces) {
 
-    private static int parseCastlingRights(String castlingRights, long whiteRooks, long blackRooks, int whiteKing, int blackKing) {
-        if (castlingRights.length() > 4) {
-            throw new IllegalArgumentException("Invalid castling rights! " + castlingRights);
+        String[] files = boardString.split("/");
+        List<List<String>> rankFileHash = Arrays.stream(files)
+                .map(file -> Arrays.stream(file.split(""))
+                        .flatMap(FEN::parseSquare)
+                        .toList())
+                .collect(Collectors.toList());
+        Collections.reverse(rankFileHash);
+
+        for (int rankIndex = 0; rankIndex < rankFileHash.size(); rankIndex++) {
+            List<String> rank = rankFileHash.get(rankIndex);
+            for (int fileIndex = 0; fileIndex < rank.size(); fileIndex++) {
+                int square = Square.of(rankIndex, fileIndex);
+                String squareValue = rank.get(fileIndex);
+                parsePiece(squareValue).ifPresent(piece -> {
+                    boolean isWhite = Character.isUpperCase(squareValue.charAt(0));
+                    updatePiecePosition(bitboards, pieces, square, piece, isWhite);
+                });
+            }
         }
+
+    }
+
+    private static int parseEnPassantFile(String epSquare) {
+        return epSquare.equals("-") ? -1 : File.of(Square.fromNotation(epSquare));
+    }
+
+    private static List<String> parseKings(String boardString) {
+        return boardString.chars()
+                .mapToObj(i -> (char) i)
+                .filter(i -> i == 'K' || i == 'k')
+                .map(String::valueOf)
+                .toList();
+    }
+
+    private static Stream<String> parseSquare(String square) {
+        if (square.length() != 1)
+            throw new InvalidFenException("Illegal square char! " + square);
+
+        boolean isLetter = Character.isLetter(square.charAt(0));
+        return isLetter ? Stream.of(square) : IntStream.range(0, Integer.parseInt(square)).mapToObj(i -> "x");
+    }
+
+    private static int parseCastlingRights(String castlingRights, long[] bitboards) {
+
+        long whiteRooks = bitboards[Piece.ROOK.index()] & bitboards[Piece.WHITE_PIECES];
+        long blackRooks = bitboards[Piece.ROOK.index()] & bitboards[Piece.BLACK_PIECES];
+        int whiteKing = Bits.next(bitboards[Piece.KING.index()] & bitboards[Piece.WHITE_PIECES]);
+        int blackKing = Bits.next(bitboards[Piece.KING.index()] & bitboards[Piece.BLACK_PIECES]);
+
         int rights = Castling.empty();
         for (int i = 0; i < castlingRights.length(); i++) {
             char right = castlingRights.charAt(i);
@@ -230,6 +287,40 @@ public record FEN(String value) {
         return rights;
     }
 
+    private static int findRook(long rooks, boolean white, boolean kingside) {
+
+        long firstRank = white ? Bits.Rank.FIRST : Bits.Rank.EIGHTH;
+        long firstRankRooks = rooks & firstRank;
+
+        if (Bits.count(firstRankRooks) == 0)
+            throw new InvalidFenException("Castling rights with no rooks on the first rank!");
+
+        if (Bits.count(firstRankRooks) == 1)
+            return Bits.next(firstRankRooks);
+
+        List<Integer> squares = Arrays.stream(Bits.collect(firstRankRooks))
+                .boxed()
+                .sorted(Comparator.comparing(File::of))
+                .toList();
+
+        return kingside ? squares.get(squares.size() - 1) : squares.get(0);
+
+    }
+
+    private static Optional<Piece> parsePiece(String code) {
+        return Arrays.stream(Piece.values())
+                .filter(piece -> piece.code().equalsIgnoreCase(code))
+                .findAny();
+    }
+
+    private static void updatePiecePosition(long[] bbs, Piece[] pcs, int square, Piece piece, boolean white) {
+        int pieceIndex = piece.index();
+        int colourIndex = Piece.COUNT + Colour.index(white);
+        bbs[pieceIndex] |= Bits.of(square);
+        bbs[colourIndex] |= Bits.of(square);
+        pcs[square] = piece;
+    }
+
     private static String toCastlingRights(int rights) {
         if (rights == Castling.empty()) {
             return "-";
@@ -254,24 +345,12 @@ public record FEN(String value) {
         return rightsString;
     }
 
-    private static int parseEnPassantFile(String enPassantSquare) {
-        if (enPassantSquare.equals("-")) {
-            return -1;
-        }
-        int square = Square.fromNotation(enPassantSquare);
-        return File.of(square);
-    }
-
     private static String toEnPassantSquare(int enPassantFile, boolean white) {
         int rank = white ? 2 : 5;
         if (enPassantFile == -1) {
             return "-";
         }
         return Square.toNotation(Square.of(rank, enPassantFile));
-    }
-
-    private static int parseFiftyMoveCounter(String fiftyMoveCounter) {
-        return Character.isDigit(fiftyMoveCounter.charAt(0)) ? Integer.parseInt(fiftyMoveCounter) : 0;
     }
 
     private static String toFiftyMoveCounter(int fiftyMoveCounter) {
@@ -282,75 +361,32 @@ public record FEN(String value) {
         return Integer.toString(1 + (ply / 2));
     }
 
-    private static Stream<String> parseSquare(String square) {
-        if (square.length() != 1)
-            throw new InvalidFenException("Illegal square char! " + square);
-
-        boolean isLetter = Character.isLetter(square.charAt(0));
-        return isLetter ? Stream.of(square) : IntStream.range(0, Integer.parseInt(square)).mapToObj(i -> "x");
+    private static String toSideToMove(boolean sideToMove) {
+        return sideToMove ? "w" : "b";
     }
 
-    public static Piece[] calculatePieceList(Board board) {
-
-        Piece[] pieceList = new Piece[Square.COUNT];
-        for (int square = 0; square < Square.COUNT; square++) {
-            long squareMask = Bits.of(square);
-            if ((squareMask & board.getPawns()) != 0)           pieceList[square] = Piece.PAWN;
-            else if ((squareMask & board.getKnights()) != 0)    pieceList[square] = Piece.KNIGHT;
-            else if ((squareMask & board.getBishops()) != 0)    pieceList[square] = Piece.BISHOP;
-            else if ((squareMask & board.getRooks()) != 0)      pieceList[square] = Piece.ROOK;
-            else if ((squareMask & board.getQueens()) != 0)     pieceList[square] = Piece.QUEEN;
-            else if ((squareMask & board.getKings()) != 0)      pieceList[square] = Piece.KING;
-        }
-        return pieceList;
-
-    }
-
-    private static int findRook(long rooks, boolean white, boolean kingside) {
-
-        long firstRank = white ? Bits.Rank.FIRST : Bits.Rank.EIGHTH;
-        long firstRankRooks = rooks & firstRank;
-
-        if (Bits.count(firstRankRooks) == 0)
-            throw new InvalidFenException("Castling rights with no rooks on the first rank!");
-
-        if (Bits.count(firstRankRooks) == 1)
-            return Bits.next(firstRankRooks);
-
-        List<Integer> squares = Arrays.stream(Bits.collect(firstRankRooks))
-                .boxed()
-                .sorted(Comparator.comparing(File::of))
-                .toList();
-
-        return kingside ? squares.get(squares.size() - 1) : squares.get(0);
-
-    }
-
-    private static Optional<Piece> parsePiece(String code) {
-        code = code.toLowerCase();
-        return switch (code) {
-            case "p" -> Optional.of(Piece.PAWN);
-            case "n" -> Optional.of(Piece.KNIGHT);
-            case "b" -> Optional.of(Piece.BISHOP);
-            case "r" -> Optional.of(Piece.ROOK);
-            case "q" -> Optional.of(Piece.QUEEN);
-            case "k" -> Optional.of(Piece.KING);
-            default -> Optional.empty();
-        };
-    }
-
-    private static void updateBitboards(long[] bbs, int square, Piece piece, boolean white) {
-        int pieceIndex = piece.index();
-        int colourIndex = Piece.COUNT + Colour.index(white);
-        bbs[pieceIndex] |= Bits.of(square);
-        bbs[colourIndex] |= Bits.of(square);
+    private static class FenPatterns {
+        public static final Pattern BOARD       = Pattern.compile("^([rnbqkpRNBQKP1-8]+/){7}[rnbqkpRNBQKP1-8]+$");
+        public static final Pattern RANK        = Pattern.compile(".*\\d{2,}.*");
+        public static final Pattern TURN        = Pattern.compile("^[wb]$");
+        public static final Pattern CASTLE      = Pattern.compile("^(-|[A-HKQ]*[a-hkq]*)$");
+        public static final Pattern EN_PASSANT  = Pattern.compile("^(-|[a-h][36])$");
+        public static final Pattern HALF_MOVE   = Pattern.compile("^\\d+$");
+        public static final Pattern FULL_MOVE   = Pattern.compile("^\\d+$");
     }
 
     public static class InvalidFenException extends RuntimeException {
+        public InvalidFenException(String message, String fen) {
+            super(String.format("Invalid FEN %s: %s", fen, message));
+        }
+
         public InvalidFenException(String message) {
-            super("Invalid FEN: " + message);
+            super(String.format("Invalid FEN: %s", message));
+        }
+
+        public InvalidFenException(String fen, Throwable cause) {
+            super(String.format("Invalid FEN %s: %s", fen, cause.getMessage()), cause);
         }
     }
-
 
 }
