+++ src/test/java/com/kelseyde/calvin/board/CheckTest.java
@@ -11,7 +11,7 @@ public class CheckTest {
     @Test
     public void checkBlocksOtherMoves() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e5"));
         board.makeMove(TestUtils.getLegalMove(board, "d1", "h5"));
@@ -29,7 +29,7 @@ public class CheckTest {
     @Test
     public void cannotMovePinnedPawn() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e5"));
         board.makeMove(TestUtils.getLegalMove(board, "d1", "h5"));
@@ -43,7 +43,7 @@ public class CheckTest {
     @Test
     public void cannotEnPassantWithPinnedPawn() {
 
-        Board board = FEN.toBoard("rnb1kbnr/ppp1qppp/8/3pP3/3p4/5N2/PPP2PPP/RNBQKB1R w KQkq - 0 1");
+        Board board = FEN.parse("rnb1kbnr/ppp1qppp/8/3pP3/3p4/5N2/PPP2PPP/RNBQKB1R w KQkq - 0 1").toBoard();
 
         // black tries to en-passant with pinned e-pawn
         Assertions.assertThrows(IllegalMoveException.class,
@@ -53,7 +53,7 @@ public class CheckTest {
 
     @Test
     public void cannotMovePinnedKnight() {
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "f7", "f5"));
         board.makeMove(TestUtils.getLegalMove(board, "e4", "f5"));
@@ -71,7 +71,7 @@ public class CheckTest {
 
     @Test
     public void cannotMovePinnedBishop() {
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "f7", "f5"));
         board.makeMove(TestUtils.getLegalMove(board, "e4", "f5"));
@@ -90,7 +90,7 @@ public class CheckTest {
     @Test
     public void cannotMoveFromCheckIntoAnotherCheck() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e5"));
         board.makeMove(TestUtils.getLegalMove(board, "d1", "h5"));
@@ -108,7 +108,7 @@ public class CheckTest {
     @Test
     public void canCaptureUnprotectedCheckingPiece() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e5"));
         board.makeMove(TestUtils.getLegalMove(board, "d1", "h5"));
@@ -124,7 +124,7 @@ public class CheckTest {
     @Test
     public void cannotCaptureProtectedCheckingPieceWithKing() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e5"));
         board.makeMove(TestUtils.getLegalMove(board, "f1", "c4"));
@@ -144,7 +144,7 @@ public class CheckTest {
     @Test
     public void canCaptureProtectedCheckingPieceWithOtherPiece() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e5"));
         board.makeMove(TestUtils.getLegalMove(board, "f1", "c4"));
@@ -163,7 +163,7 @@ public class CheckTest {
     @Test
     public void cannotCastleOutOfCheck() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e6"));
         board.makeMove(TestUtils.getLegalMove(board, "d2", "d4"));
@@ -187,7 +187,7 @@ public class CheckTest {
     @Test
     public void cannotCastleThroughCheck() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         board.makeMove(TestUtils.getLegalMove(board, "e7", "e5"));
         board.makeMove(TestUtils.getLegalMove(board, "f2", "f4"));
@@ -207,7 +207,7 @@ public class CheckTest {
     public void cannotCastleQueensideThroughKnightCheck() {
 
         String fen = "r3k2r/p1ppqpb1/bnN1pnp1/3P4/1p2P3/2N2Q1p/PPPBBPPP/R3K2R b KQkq - 1 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         // try to castle through the bishop check
         Assertions.assertThrows(IllegalMoveException.class,
@@ -219,7 +219,7 @@ public class CheckTest {
     public void cannotStepKingBackOnRayOfCheckingOrthogonalSlider() {
 
         String fen = "8/8/8/2k5/2r2K2/8/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         // try to step away on same rank as rook
         Assertions.assertThrows(IllegalMoveException.class,
@@ -231,7 +231,7 @@ public class CheckTest {
     public void cannotStepKingBackOnRayOfCheckingDiagonalSlider() {
 
         String fen = "8/8/3b4/2k5/5K2/8/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         // try to step away on same diagonal as bishop
         Assertions.assertThrows(IllegalMoveException.class,
@@ -244,7 +244,7 @@ public class CheckTest {
 
         // fen = en passant funhouse fen
         String fen = "8/2p5/3p4/KP5r/1R3p1k/8/4P1P1/8 w - - ";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         board.makeMove(TestUtils.getLegalMove(board, "e2", "e4"));
         Assertions.assertThrows(IllegalMoveException.class,
@@ -257,7 +257,7 @@ public class CheckTest {
 
         String fen = "8/2p5/3p4/1P4r1/1K3p1k/8/4P1P1/1R6 b - - 3 2";
 
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         board.makeMove(TestUtils.getLegalMove(board, "c7", "c5"));
         board.makeMove(TestUtils.getLegalMove(board, "b5", "c6"));
@@ -268,7 +268,7 @@ public class CheckTest {
     public void testDiagonallyPinnedPawnCannotMove() {
 
         String fen = "6k1/8/8/8/6b1/8/4P3/3K4 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
         Assertions.assertThrows(IllegalMoveException.class,
                 () -> TestUtils.getLegalMove(board, "e2", "e3"));
         Assertions.assertThrows(IllegalMoveException.class,
@@ -279,7 +279,7 @@ public class CheckTest {
     public void testPinnedOrthogonalSliderCanStepForwardInThePinRay() {
 
         String fen = "6k1/8/3r4/8/8/8/3R4/3K4 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
         TestUtils.getLegalMove(board, "d2", "d3");
         TestUtils.getLegalMove(board, "d2", "d6");
     }
@@ -288,7 +288,7 @@ public class CheckTest {
     public void testPinnedOrthogonalSliderCanStepBackwardInThePinRay() {
 
         String fen = "6k1/8/3r4/3R4/8/8/8/3K4 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
         TestUtils.getLegalMove(board, "d5", "d4");
         TestUtils.getLegalMove(board, "d5", "d3");
     }
@@ -297,7 +297,7 @@ public class CheckTest {
     public void testPinnedDiagonalSliderCanStepForwardInThePinRay() {
 
         String fen = "6k1/8/8/7q/8/5B2/8/3K4 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
         TestUtils.getLegalMove(board, "f3", "g4");
         TestUtils.getLegalMove(board, "f3", "h5");
     }
@@ -306,7 +306,7 @@ public class CheckTest {
     public void testPinnedDiagonalSliderCanStepBackwardInThePinRay() {
 
         String fen = "6k1/8/8/7q/6B1/8/8/3K4 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
         TestUtils.getLegalMove(board, "g4", "f3");
         TestUtils.getLegalMove(board, "g4", "e2");
     }
@@ -314,7 +314,7 @@ public class CheckTest {
     @Test
     public void canBlockCheckWithPawnPush() {
 
-        Board board = FEN.toBoard("rnbqkbnr/ppp1pppp/3p4/8/Q7/2P5/PP1PPPPP/RNB1KBNR b KQkq - 1 2");
+        Board board = FEN.parse("rnbqkbnr/ppp1pppp/3p4/8/Q7/2P5/PP1PPPPP/RNB1KBNR b KQkq - 1 2").toBoard();
 
         TestUtils.getLegalMove(board, "b7", "b5");
 
@@ -323,7 +323,7 @@ public class CheckTest {
     @Test
     public void kingCannotMoveIntoRookCheck() {
 
-        Board board = FEN.toBoard("r5k1/5b2/q2p4/p2nn1P1/2p5/P3P3/1PB2PK1/2BR3R b - - 2 42");
+        Board board = FEN.parse("r5k1/5b2/q2p4/p2nn1P1/2p5/P3P3/1PB2PK1/2BR3R b - - 2 42").toBoard();
 
         Assertions.assertThrows(IllegalMoveException.class, () ->
                 TestUtils.getLegalMove(board, Move.fromUCI("g8h8")));
