+++ src/test/java/com/kelseyde/calvin/board/KeyTest.java
@@ -12,8 +12,8 @@ public class KeyTest {
     @Test
     public void testSamePositionGeneratesSameKey() {
 
-        Board board1 = Board.from(FEN.STARTPOS);
-        Board board2 = Board.from(FEN.STARTPOS);
+        Board board1 = FEN.startpos().toBoard();
+        Board board2 = FEN.startpos().toBoard();
         Assertions.assertEquals(board1.getState().getKey(), board2.getState().getKey());
 
         Move e4 = new Move(12, 28, Move.PAWN_DOUBLE_MOVE_FLAG);
@@ -26,7 +26,7 @@ public class KeyTest {
         board2.unmakeMove();
 
         Assertions.assertEquals(board1.getState().getKey(), board2.getState().getKey());
-        Assertions.assertEquals(board1.getState().getKey(), Board.from(FEN.STARTPOS).getState().getKey());
+        Assertions.assertEquals(board1.getState().getKey(), FEN.startpos().toBoard().getState().getKey());
 
     }
 
@@ -35,7 +35,7 @@ public class KeyTest {
 
         String fen = "k6K/1pp2P1P/p1p5/P7/8/8/5r2/2R5 w - - 1 51";
 
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
         long firstZobrist1 = board.getState().getKey();
 
         board.makeMove(Move.fromUCI("h8g8"));
@@ -63,7 +63,7 @@ public class KeyTest {
 
         String fen = "1rb3k1/p1q3pp/4pr2/5p2/2pP4/1PQ3P1/4PPBP/2R1K2R b K - 0 21";
 
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
         long z1 = board.getState().getKey();
         board.makeMove(Move.fromUCI("b8b3"));
         long z2 = board.getState().getKey();
@@ -89,11 +89,11 @@ public class KeyTest {
         String fenBeforeMove = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";
         String fenAfterMove = "rnbqkbnr/pppppppp/8/8/4P3/8/PPPP1PPP/RNBQKBNR b KQkq - 0 1";
 
-        Board board1 = FEN.toBoard(fenBeforeMove);
+        Board board1 = FEN.parse(fenBeforeMove).toBoard();
         board1.makeMove(Move.fromUCI("e2e4"));
         long zobrist1 = board1.getState().getKey();
 
-        Board board2 = FEN.toBoard(fenAfterMove);
+        Board board2 = FEN.parse(fenAfterMove).toBoard();
         long zobrist2 = board2.getState().getKey();
 
         Assertions.assertEquals(zobrist1, zobrist2);
@@ -106,11 +106,11 @@ public class KeyTest {
         String fenBeforeMove = "rnbqkbnr/ppp1pppp/8/3p4/4P3/8/PPPP1PPP/RNBQKBNR w KQkq - 0 2";
         String fenAfterMove = "rnbqkbnr/ppp1pppp/8/3P4/8/8/PPPP1PPP/RNBQKBNR b KQkq - 0 2";
 
-        Board board1 = FEN.toBoard(fenBeforeMove);
+        Board board1 = FEN.parse(fenBeforeMove).toBoard();
         board1.makeMove(Move.fromUCI("e4d5"));
         long zobrist1 = board1.getState().getKey();
 
-        Board board2 = FEN.toBoard(fenAfterMove);
+        Board board2 = FEN.parse(fenAfterMove).toBoard();
         long zobrist2 = board2.getState().getKey();
 
         Assertions.assertEquals(zobrist1, zobrist2);
@@ -123,11 +123,11 @@ public class KeyTest {
         String fenBeforeMove = "rnbqkbnr/ppp1pppp/8/3p4/4P3/8/PPPP1PPP/RNBQKBNR w KQkq - 0 2";
         String fenAfterMove = "rnbqkbnr/ppp1pppp/8/3p4/4P3/8/PPPPKPPP/RNBQ1BNR b kq - 1 2";
 
-        Board board1 = FEN.toBoard(fenBeforeMove);
+        Board board1 = FEN.parse(fenBeforeMove).toBoard();
         board1.makeMove(Move.fromUCI("e1e2"));
         long zobrist1 = board1.getState().getKey();
 
-        Board board2 = FEN.toBoard(fenAfterMove);
+        Board board2 = FEN.parse(fenAfterMove).toBoard();
         long zobrist2 = board2.getState().getKey();
 
         Assertions.assertEquals(zobrist1, zobrist2);
@@ -140,11 +140,11 @@ public class KeyTest {
         String fenBeforeCapture = "rnbqkb1r/ppp1pppp/5n2/3pP3/8/8/PPPP1PPP/RNBQKBNR w KQkq d6 0 3";
         String fenAfterCapture = "rnbqkb1r/ppp1pppp/3P1n2/8/8/8/PPPP1PPP/RNBQKBNR b KQkq - 0 3";
 
-        Board board1 = FEN.toBoard(fenBeforeCapture);
+        Board board1 = FEN.parse(fenBeforeCapture).toBoard();
         board1.makeMove(Move.fromUCI("e5d6", Move.EN_PASSANT_FLAG));
         long zobrist1 = board1.getState().getKey();
 
-        Board board2 = FEN.toBoard(fenAfterCapture);
+        Board board2 = FEN.parse(fenAfterCapture).toBoard();
         long zobrist2 = board2.getState().getKey();
 
         Assertions.assertEquals(zobrist1, zobrist2);
@@ -156,11 +156,11 @@ public class KeyTest {
         String fenBeforeCastle = "r1bqk1nr/pppp1ppp/2n5/2b1p3/2B1P3/5N2/PPPP1PPP/RNBQK2R w KQkq - 4 4";
         String fenAfterCastle = "r1bqk1nr/pppp1ppp/2n5/2b1p3/2B1P3/5N2/PPPP1PPP/RNBQ1RK1 b kq - 5 4";
 
-        Board board1 = FEN.toBoard(fenBeforeCastle);
+        Board board1 = FEN.parse(fenBeforeCastle).toBoard();
         board1.makeMove(Move.fromUCI("e1g1", Move.CASTLE_FLAG));
         long zobrist1 = board1.getState().getKey();
 
-        Board board2 = FEN.toBoard(fenAfterCastle);
+        Board board2 = FEN.parse(fenAfterCastle).toBoard();
         long zobrist2 = board2.getState().getKey();
 
         Assertions.assertEquals(zobrist1, zobrist2);
@@ -173,11 +173,11 @@ public class KeyTest {
         String fenBeforeCastle = "rnbqkb1r/pP3ppp/5n2/4p3/8/8/PPPP1PPP/RNBQKBNR w - - 0 7";
         String fenAfterCastle = "Qnbqkb1r/p4ppp/5n2/4p3/8/8/PPPP1PPP/RNBQKBNR b - - 0 7";
 
-        Board board1 = FEN.toBoard(fenBeforeCastle);
+        Board board1 = FEN.parse(fenBeforeCastle).toBoard();
         board1.makeMove(Move.fromUCI("b7a8", Move.PROMOTE_TO_QUEEN_FLAG));
         long zobrist1 = board1.getState().getKey();
 
-        Board board2 = FEN.toBoard(fenAfterCastle);
+        Board board2 = FEN.parse(fenAfterCastle).toBoard();
         long zobrist2 = board2.getState().getKey();
 
         Assertions.assertEquals(zobrist1, zobrist2);
@@ -190,11 +190,11 @@ public class KeyTest {
         String fenBeforePromotion = "rnb1kb1r/pP2pppp/5n2/8/8/3q4/PPPP1PPP/RNBQKBNR w KQkq - 1 5";
         String fenAfterPromotion = "Qnb1kb1r/p3pppp/5n2/8/8/3q4/PPPP1PPP/RNBQKBNR b KQk - 0 5";
 
-        Board board1 = FEN.toBoard(fenBeforePromotion);
+        Board board1 = FEN.parse(fenBeforePromotion).toBoard();
         board1.makeMove(Move.fromUCI("b7a8q", Move.PROMOTE_TO_QUEEN_FLAG));
         long zobrist1 = board1.getState().getPawnKey();
 
-        Board board2 = FEN.toBoard(fenAfterPromotion);
+        Board board2 = FEN.parse(fenAfterPromotion).toBoard();
         long zobrist2 = board2.getState().getPawnKey();
 
         Assertions.assertEquals(zobrist1, zobrist2);
@@ -204,7 +204,7 @@ public class KeyTest {
     @Test
     public void testNonPawnNotUpdatedOnPawnMove() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         long[] keys = board.nonPawnKeys();
 
         board.makeMove(Move.fromUCI("e2e4"));
@@ -221,7 +221,7 @@ public class KeyTest {
     @Test
     public void testNonPawnCorrectSideUpdatedWhite() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         long[] keys = Arrays.copyOf(board.nonPawnKeys(), board.nonPawnKeys().length);
 
         board.makeMove(Move.fromUCI("g1f3"));
@@ -236,7 +236,7 @@ public class KeyTest {
     @Test
     public void testNonPawnCorrectSideUpdatedBlack() {
 
-        Board board = Board.from(FEN.STARTPOS);
+        Board board = FEN.startpos().toBoard();
         long[] keys = Arrays.copyOf(board.nonPawnKeys(), board.nonPawnKeys().length);
 
         board.makeMove(Move.fromUCI("e2e4", Move.PAWN_DOUBLE_MOVE_FLAG));
