+++ src/test/java/com/kelseyde/calvin/movegen/PinCalculatorTest.java
@@ -13,7 +13,7 @@ public class PinCalculatorTest {
     public void testNoSlidersAndNoPins() {
 
         String fen = "4nk2/5p2/8/8/8/2PP4/2K1N3/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0L, movegen.getPinMask());
@@ -27,7 +27,7 @@ public class PinCalculatorTest {
     public void testDiagonalSlidersButNoPins() {
 
         String fen = "4qkb1/5pb1/8/8/8/2PP4/2KBB3/2Q5 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0L, movegen.getPinMask());
@@ -41,7 +41,7 @@ public class PinCalculatorTest {
     public void testOrthogonalSlidersButNoPins() {
 
         String fen = "4qkr1/5pr1/3r4/8/8/2PP2R1/2KRR3/2Q5 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0L, movegen.getPinMask());
@@ -55,7 +55,7 @@ public class PinCalculatorTest {
     public void testOrthogonalCheckIsNotAPin() {
 
         String fen = "8/8/3rk3/8/8/3K4/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0L, movegen.getPinMask());
@@ -69,7 +69,7 @@ public class PinCalculatorTest {
     public void testDiagonalCheckIsNotAPin() {
 
         String fen = "8/8/4k3/5b2/8/3K4/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0L, movegen.getPinMask());
@@ -83,7 +83,7 @@ public class PinCalculatorTest {
     public void testCannotOrthogonallyPinThroughSameSidePiece() {
 
         String fen = "5k2/4q3/4p3/8/8/8/8/4K3 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0L, movegen.getPinMask());
@@ -96,7 +96,7 @@ public class PinCalculatorTest {
     public void testSimplePositiveFileOrthogonalPin() {
 
         String fen = "4k3/4r3/8/8/8/8/4P3/4K3 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(1L << 12, movegen.getPinMask());
@@ -110,7 +110,7 @@ public class PinCalculatorTest {
     public void testSimpleNegativeFileOrthogonalPin() {
 
         String fen = "4k3/4b3/8/8/8/8/4R3/4K3 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0L, movegen.getPinMask());
@@ -124,7 +124,7 @@ public class PinCalculatorTest {
     public void testFileOrthogonalTwoWayPin() {
 
         String fen = "8/3k4/3q4/8/8/3R4/3K4/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(1L << 19, movegen.getPinMask());
@@ -138,7 +138,7 @@ public class PinCalculatorTest {
     public void testFileOrthogonalDoubleFriendlyPinEqualsNoPin() {
 
         String fen = "8/3k4/3q4/3q4/3R4/3R4/3K4/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0, movegen.getPinMask());
@@ -152,7 +152,7 @@ public class PinCalculatorTest {
     public void testFileOrthogonalDoubleOpponentPinIsStillPin() {
 
         String fen = "4k3/4q3/4r3/8/4N3/4K3/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(1L << 28, movegen.getPinMask());
@@ -166,7 +166,7 @@ public class PinCalculatorTest {
     public void testSimplePositiveRankOrthogonalPin() {
 
         String fen = "KR4nk/8/8/8/8/8/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0, movegen.getPinMask());
@@ -180,7 +180,7 @@ public class PinCalculatorTest {
     public void testSimpleNegativeRankOrthogonalPin() {
 
         String fen = "KN4qk/8/8/8/8/8/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(1L << 57, movegen.getPinMask());
@@ -194,7 +194,7 @@ public class PinCalculatorTest {
     public void testSimplePositiveDiagonalPin() {
 
         String fen = "K1n5/2q5/nqr1B3/5r2/6k1/8/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0, movegen.getPinMask());
@@ -208,7 +208,7 @@ public class PinCalculatorTest {
     public void testSimpleNegativeDiagonalPin() {
 
         String fen = "8/8/8/8/1k6/2q5/3QK3/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0, movegen.getPinMask());
@@ -222,7 +222,7 @@ public class PinCalculatorTest {
     public void testDiagonalDoubleFriendlyPinEqualsNoPin() {
 
         String fen = "7k/8/8/1b6/2N5/3R4/3QK3/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0, movegen.getPinMask());
@@ -236,7 +236,7 @@ public class PinCalculatorTest {
     public void testSimplePositiveAntiDiagonalPin() {
 
         String fen = "7k/6b1/8/8/8/8/1B6/1KRRRRR1 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(0, movegen.getPinMask());
@@ -250,7 +250,7 @@ public class PinCalculatorTest {
     public void testSimpleNegativeAntiDiagonalPin() {
 
         String fen = "5q2/4Q3/3K4/7r/7r/7r/7r/7k w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(1L << 52, movegen.getPinMask());
@@ -264,7 +264,7 @@ public class PinCalculatorTest {
     public void testTwoWayAntiDiagonalPin() {
 
         String fen = "7K/8/8/4B3/3b4/8/8/k7 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(1L << 36, movegen.getPinMask());
@@ -278,7 +278,7 @@ public class PinCalculatorTest {
     public void testAntiDiagonalDoubleFriendlyPinEqualsNoPin() {
 
         String fen = "7K/8/8/4B3/3b4/2b5/8/k7 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         movegen.calculatePins(board, true);
         Assertions.assertEquals(1L << 36, movegen.getPinMask());
@@ -292,7 +292,7 @@ public class PinCalculatorTest {
     public void testBothSidesOrthogonalRankPin() {
 
         String fen = "3r4/8/8/5k2/8/8/8/r1NKN2r w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         long expectedPinMask = 0L;
         expectedPinMask |= 1L << 2;
@@ -310,7 +310,7 @@ public class PinCalculatorTest {
     public void testBothSidesOrthogonalFilePin() {
 
         String fen = "K7/8/4R3/4p3/4k3/4p3/4Q3/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         long expectedPinMask = 0L;
         expectedPinMask |= 1L << 20;
@@ -328,7 +328,7 @@ public class PinCalculatorTest {
     public void testBothSidesAntiDiagonalPin() {
 
         String fen = "K7/8/6B1/5p2/4k3/3p4/2B5/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         long expectedPinMask = 0L;
         expectedPinMask |= 1L << 19;
@@ -346,7 +346,7 @@ public class PinCalculatorTest {
     public void testBothSidesDiagonalPin() {
 
         String fen = "K7/8/2B5/3q1p2/4k3/3p1r2/6B1/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         long expectedPinMask = 0L;
         expectedPinMask |= 1L << 35;
@@ -363,7 +363,7 @@ public class PinCalculatorTest {
     public void testCrazyPinsEverywhere() {
 
         String fen = "1b1R1q2/2PnP3/rNNK4/4NR2/4pbp1/5k2/4bnp1/3B1R1R w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         long expectedWhitePinMask = 0L;
         expectedWhitePinMask |= 1L << 50;
