+++ src/test/java/com/kelseyde/calvin/movegen/quiescent/QuiescentTest.java
@@ -18,7 +18,7 @@ public class QuiescentTest {
 
         // Includes a pinned knight and a friendly blocker
         String fen = "1k1r1N1n/pppb3p/P2b4/4N3/1N6/6K1/8/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         List<Move> moves = moveGenerator.generateMoves(board, MoveGenerator.MoveFilter.CAPTURES_ONLY);
         List<Move> expected = List.of(
@@ -43,7 +43,7 @@ public class QuiescentTest {
     public void testGenerateBishopMoves() {
 
         String fen = "6k1/4bpp1/7p/8/p6b/1B4B1/2B2K2/8 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         List<Move> moves = moveGenerator.generateMoves(board, MoveGenerator.MoveFilter.CAPTURES_ONLY);
         List<Move> expected = List.of(
@@ -68,7 +68,7 @@ public class QuiescentTest {
     public void testGenerateQueenMoves() {
 
         String fen = "8/8/1KQr4/2Q5/6p1/6pk/2Q4N/4r3 w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         List<Move> moves = moveGenerator.generateMoves(board, MoveGenerator.MoveFilter.CAPTURES_ONLY);
         List<Move> expected = List.of(
@@ -95,7 +95,7 @@ public class QuiescentTest {
     public void testPawnAndPromotionMoves() {
 
         String fen = "3k4/8/8/8/6pP/1p2p3/3B1p2/2K3R1 b - h3 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         List<Move> moves = moveGenerator.generateMoves(board, MoveGenerator.MoveFilter.CAPTURES_ONLY);
         List<Move> expected = List.of(
@@ -134,7 +134,7 @@ public class QuiescentTest {
     public void testRookMoves() {
 
         String fen = "1bR2N1b/R3nk2/5p2/8/8/8/2K3R1/4R2R w - - 0 1";
-        Board board = FEN.toBoard(fen);
+        Board board = FEN.parse(fen).toBoard();
 
         List<Move> moves = moveGenerator.generateMoves(board, MoveGenerator.MoveFilter.CAPTURES_ONLY);
         List<Move> expected = List.of(
